compilation + coding-style -- start menu disabled | todo: manage input + sfView to focus player + add player
